# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/token": {
      "post": {
        "tags": ["auth"],
        "summary": "Login For Access Token",
        "operationId": "login_for_access_token_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Token" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/change_password": {
      "post": {
        "tags": ["auth"],
        "summary": "Change Password",
        "operationId": "change_password_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/fetch_items": {
      "get": {
        "tags": ["core"],
        "summary": "Fetch Items",
        "operationId": "fetch_items_fetch_items_get",
        "parameters": [
          {
            "name": "only_wish_list",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Only Wish List"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Sortorder"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Limit"
            }
          },
          {
            "name": "searchQuery",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Searchquery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FetchItemsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_ean_to_list/": {
      "post": {
        "tags": ["core"],
        "summary": "Add Ean To List",
        "operationId": "add_ean_to_list_add_ean_to_list__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddEanRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_fetched_items": {
      "post": {
        "tags": ["core"],
        "summary": "Add Fetched Items",
        "operationId": "add_fetched_items_add_fetched_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddFetchedItems" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/classify_items_against_pantry": {
      "get": {
        "tags": ["core"],
        "summary": "Classify Items Against Pantry",
        "operationId": "classify_items_against_pantry_classify_items_against_pantry_get",
        "parameters": [
          {
            "name": "itemsWished",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Itemswished" }
          },
          {
            "name": "autoAddItems",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "default": true,
              "title": "Autoadditems"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddEanRequest": {
        "properties": {
          "ean": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Ean"
          },
          "item_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Item Name"
          },
          "count": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Count",
            "default": 1
          },
          "wish_list": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Wish List"
          },
          "quantity_data": {
            "anyOf": [
              { "$ref": "#/components/schemas/QuantityInfo" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "title": "AddEanRequest",
        "description": "Request model for add_ean_to_list endpoint"
      },
      "AddFetchedItems": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/FetchedSingleItem" },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "AddFetchedItems"
      },
...
      "Body_login_for_access_token_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              { "type": "string", "pattern": "^password$" },
              { "type": "null" }
            ],
            "title": "Grant Type"
          },
          "username": { "type": "string", "title": "Username" },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": { "type": "string", "title": "Scope", "default": "" },
          "client_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "Body_login_for_access_token_token_post"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": { "type": "string", "title": "Current Password" },
          "new_password": { "type": "string", "title": "New Password" }
        },
        "type": "object",
        "required": ["current_password", "new_password"],
        "title": "ChangePasswordRequest"
      },
...
      "FetchItemsResponse": {
        "properties": {
          "status": { "type": "integer", "const": 200, "title": "Status" },
          "items": {
            "items": { "$ref": "#/components/schemas/ItemsFetched" },
            "type": "array",
            "title": "Items"
          },
          "distinct_items": { "type": "integer", "title": "Distinct Items" },
          "accumulated_count": {
            "type": "integer",
            "title": "Accumulated Count"
          }
        },
        "type": "object",
        "required": ["status", "items", "distinct_items", "accumulated_count"],
        "title": "FetchItemsResponse"
      },
      "FetchedSingleItem": {
        "properties": {
          "userId": { "type": "string", "title": "Userid" },
          "ean": { "type": "string", "title": "Ean" },
          "item_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Item Name"
          },
          "per_date_count": {
            "items": { "type": "integer" },
            "type": "array",
            "title": "Per Date Count"
          },
          "perish_dates": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Perish Dates"
          },
          "isWished": { "type": "string", "title": "Iswished" }
        },
        "type": "object",
        "required": [
          "userId",
          "ean",
          "per_date_count",
          "perish_dates",
          "isWished"
        ],
        "title": "FetchedSingleItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemsFetched": {
        "properties": {
          "ean": { "type": "string", "title": "Ean" },
          "text": { "type": "string", "title": "Text" },
          "shortened_name": { "type": "string", "title": "Shortened Name" },
          "count": { "type": "integer", "title": "Count" },
          "perish_dates": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Perish Dates"
          },
          "imageUrl": { "type": "string", "title": "Imageurl" },
          "tags": { "type": "string", "title": "Tags" },
          "mappedItems": {
            "items": {},
            "type": "array",
            "title": "Mappeditems"
          },
          "quantity": {
            "anyOf": [
              { "$ref": "#/components/schemas/QuantityInfo" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": [
          "ean",
          "text",
          "shortened_name",
          "count",
          "perish_dates",
          "imageUrl",
          "tags",
          "mappedItems"
        ],
        "title": "ItemsFetched"
      },
...
      "QuantityInfo": {
        "properties": {
          "product_quantity": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Product Quantity"
          },
          "product_quantity_unit": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Quantity Unit"
          }
        },
        "type": "object",
        "required": ["product_quantity", "product_quantity_unit"],
        "title": "QuantityInfo"
      },
      "Token": {
        "properties": {
          "access_token": { "type": "string", "title": "Access Token" },
          "token_type": { "type": "string", "title": "Token Type" }
        },
        "type": "object",
        "required": ["access_token", "token_type"],
        "title": "Token"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}